pyproject.toml
[project]
name = "python-chat-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "streamlit>=1.46.1",
]
